requirements.txt
# requirements.txt — Multilingual RAG System
# Core
fastapi>=0.104.0
uvicorn>=0.24.0
sqlalchemy>=2.0.23
pydantic>=2.5.0
pydantic-settings>=2.1.0

# ML & Embeddings
sentence-transformers>=2.6.0
faiss-cpu>=1.8.0
numpy>=1.26.0
scikit-learn>=1.3.0

# Data Processing
pandas>=2.1.0
tqdm>=4.66.0

# NLP & Evaluation
nltk>=3.8.0
rouge-score>=0.1.2

# LLM
groq>=0.4.0

# Caching
cachetools>=5.3.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0

# Utilities
aiofiles>=23.2.0
python-dotenv>=1.0.0
httpx>=0.25.0

# Testing
pytest>=7.4.0